% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_preprocessing.R
\name{mod_data_preprocessing_server}
\alias{mod_data_preprocessing_server}
\title{Data-preprocessing module server.}
\usage{
mod_data_preprocessing_server(id)
}
\arguments{
\item{id}{Namespace identifier for the module.}
}
\value{
A named list containing \code{data}, \code{status}, and \code{error_message}
reactive expressions.
}
\description{
Data-preprocessing module server.
}
