% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_preprocessing.R
\name{mod_data_preprocessing_ui}
\alias{mod_data_preprocessing_ui}
\title{Data-preprocessing module UI.}
\usage{
mod_data_preprocessing_ui(id)
}
\arguments{
\item{id}{Namespace identifier for the module.}
}
\value{
A Shiny UI fragment.
}
\description{
Data-preprocessing module UI.
}
