% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlate.R
\name{plotPlate}
\alias{plotPlate}
\title{Plot Screening Plate Data}
\usage{
plotPlate(
  screenData,
  plate = "all",
  plotType = "viability",
  facetwrap = TRUE,
  outputPath = NULL,
  ncol = 2,
  nrow = 3,
  width = 16,
  height = 16,
  returnObject = FALSE
)
}
\arguments{
\item{screenData}{The data frame containing the screen data to be plotted. It must include columns for \code{experimentID}, \code{WellID}, and the relevant measurement (e.g., \code{viab_norm} for viability plots).}

\item{plate}{A list of character vectors describing a subset of plates to plot.}

\item{plotType}{A character string specifying the type of plot to generate.}

\item{facetwrap}{A logical value indicating whether to use facet wrapping for multiple plates.}

\item{outputPath}{A character string specifying the path to save the plots. If NULL, plots will not be saved to disk.}

\item{ncol}{An integer specifying the number of columns in the facet wrap layout. Default is 2.}

\item{nrow}{An integer specifying the number of rows in the facet wrap layout. Default is 3.}

\item{width}{A numeric value specifying the width of the output plot in inches. Default is 16.}

\item{height}{A numeric value specifying the height of the output plot in inches. Default is 16.}

\item{returnObject}{A logical value indicating whether to return the plot object instead of displaying it.}
}
\value{
A list of ggplot objects representing the generated plots.
}
\description{
This function generates visualizations of screening plate data, including heatmaps of viability, Z-scores, layout, and edge effects. It supports plotting multiple plates with optional facet wrapping and allows for saving the plots to disk.
}
\examples{
\dontrun{
# Example usage of plotPlate function
screenData <- readScreen("wellInput.csv", "plateInput.csv")
plotPlate(screenData, plate = "all", plotType = "viability", facetwrap = TRUE, outputPath = "plots", ncol = 2, nrow = 3, width = 16, height = 16, returnObject = FALSE)
}
}
